Update to Master

+++ app/call_recordings/call_recordings.php
@@ -42,24 +42,35 @@
 	$language = new text;
 	$text = $language->get();
 
-//get variables used to control the order
-	$order_by = check_str($_GET["order_by"]);
-	$order = check_str($_GET["order"]);
+//get the action
+	if (strlen($_REQUEST["search"]) == 0 && is_array($_POST["call_recordings"])) {
+		$call_recordings = $_POST["call_recordings"];
+		foreach($call_recordings as $row) {
+			if ($row['action'] === 'download') {
+				$action = 'download';
+				break;
+			}
+			if ($row['action'] === 'delete') {
+				$action = 'delete';
+				break;
+			}
+		}
+	}
 
-//add the search term
-	$search = strtolower(check_str($_GET["search"]));
-	if (strlen($search) > 0) {
-		$sql_search = "and (";
-		$sql_search .= "lower(call_recording_name) like '%".$search."%' ";
-		$sql_search .= "or lower(call_recording_path) like '%".$search."%' ";
-		$sql_search .= "or lower(call_direction) like '%".$search."%' ";
-		$sql_search .= "or lower(call_recording_description) like '%".$search."%' ";
-		$sql_search .= ") ";
+//download recordings
+	if (permission_exists('call_recording_download_add')) {
+		if ($action == "download") {
+			//download
+				$obj = new call_recording_downloads;
+				$obj->save($call_recordings);
+			//direct the user to the downloads
+				header("Location: ".PROJECT_PATH."/app/call_recording_downloads/call_recording_downloads.php");
+		}
 	}
 
 //delete the recordings
 	if (permission_exists('call_recording_delete')) {
-		if (is_array($_POST["call_recordings"])) {
+		if ($action === "delete") {
 			//set the array
 				$call_recordings = $_POST["call_recordings"];
 			//download
@@ -74,11 +85,31 @@
 	require_once "resources/header.php";
 	require_once "resources/paging.php";
 
+//get variables used to control the order
+	$order_by = check_str($_REQUEST["order_by"]);
+	$order = check_str($_REQUEST["order"]);
+
+//set the defaults
+	if (strlen($order_by) == 0) { 
+		$order_by = 'call_recording_date';
+		$order = 'desc';
+	}
+
+//add the search term
+	$search = strtolower(check_str($_REQUEST["search"]));
+	if (strlen($search) > 0) {
+		$sql_search = "and (";
+		$sql_search .= "lower(call_recording_name) like '%".$search."%' ";
+		$sql_search .= "or lower(call_recording_path) like '%".$search."%' ";
+		$sql_search .= "or lower(call_direction) like '%".$search."%' ";
+		$sql_search .= "or lower(call_recording_description) like '%".$search."%' ";
+		$sql_search .= ") ";
+	}
+
 //prepare to page the results
 	$sql = "select count(call_recording_uuid) as num_rows from v_call_recordings ";
 	$sql .= "where domain_uuid = '".$_SESSION["domain_uuid"]."' ";
 	$sql .= $sql_search;
-	if (strlen($order_by)> 0) { $sql .= "order by $order_by $order "; }
 	$prep_statement = $db->prepare($sql);
 	if ($prep_statement) {
 		$prep_statement->execute();
@@ -94,8 +125,8 @@
 //prepare to page the results
 	$rows_per_page = ($_SESSION['domain']['paging']['numeric'] != '') ? $_SESSION['domain']['paging']['numeric'] : 50;
 	$param = "";
-	$page = $_GET['page'];
-	if (strlen($page) == 0) { $page = 0; $_GET['page'] = 0; }
+	$page = $_REQUEST['page'];
+	if (strlen($page) == 0) { $page = 0; $_REQUEST['page'] = 0; }
 	list($paging_controls, $rows_per_page, $var3) = paging($num_rows, $param, $rows_per_page);
 	$offset = $rows_per_page * $page;
 
@@ -103,7 +134,7 @@
 	$sql = "select * from v_call_recordings ";
 	$sql .= "where domain_uuid = '".$_SESSION["domain_uuid"]."' ";
 	$sql .= $sql_search;
-	if (strlen($order_by)> 0) { $sql .= "order by $order_by $order "; }
+	$sql .= "order by $order_by $order ";
 	$sql .= "limit $rows_per_page offset $offset ";
 	$prep_statement = $db->prepare(check_sql($sql));
 	$prep_statement->execute();
@@ -120,8 +151,8 @@
 	echo "	function checkbox_toggle(item) {\n";
 	echo "		var inputs = document.getElementsByTagName(\"input\");\n";
 	echo "		for (var i = 0, max = inputs.length; i < max; i++) {\n";
-	echo "		    if (inputs[i].type === 'checkbox') {\n";
-	echo "		       	if (document.getElementById('checkbox_all').checked == true) {\n";
+	echo "			if (inputs[i].type === 'checkbox') {\n";
+	echo "				if (document.getElementById('checkbox_all').checked == true) {\n";
 	echo "				inputs[i].checked = true;\n";
 	echo "			}\n";
 	echo "				else {\n";
@@ -133,15 +164,18 @@
 	echo "</script>\n";
 
 //show the content
+	echo "<form method='post' action=''>\n";
 	echo "<table width='100%' border='0'>\n";
 	echo "	<tr>\n";
 	echo "		<td width='50%' align='left' nowrap='nowrap'><b>".$text['title-call_recordings']."</b></td>\n";
-	echo "		<form method='get' action=''>\n";
 	echo "			<td width='50%' style='vertical-align: top; text-align: right; white-space: nowrap;'>\n";
+	if (permission_exists('call_recording_download_add')) {
+		echo "				<button type='submit' class='btn btn-default' id='downloads' name=\"call_recordings[$x][action]\" alt='".$text['button-download']."' onclick=\"document.getElementById('downloads').value='download'\" value=''>".$text['button-downloads']."</span></button>\n";
+		echo "				&nbsp; &nbsp; &nbsp; ";
+	}
 	echo "				<input type='text' class='txt' style='width: 150px' name='search' id='search' value='".$search."'>\n";
 	echo "				<input type='submit' class='btn' name='submit' value='".$text['button-search']."'>\n";
 	echo "			</td>\n";
-	echo "		</form>\n";
 	echo "	</tr>\n";
 	echo "	<tr>\n";
 	echo "		<td align='left' colspan='2'>\n";
@@ -160,10 +194,9 @@
 	//echo "</style>\n";
 	echo "\n";
 
-	echo "<form method='post' action=''>\n";
 	echo "<table class='tr_hover' width='100%' border='0' cellpadding='0' cellspacing='0'>\n";
 	echo "<tr>\n";
-	echo "	<th style='width:4px;'>\n";
+	echo "	<th style='width:30px;'>\n";
 	echo "		<input type='checkbox' name='checkbox_all' id='checkbox_all' value='' onclick=\"checkbox_toggle();\">\n";
 	echo "	</th>\n";
 	echo th_order_by('call_recording_name', $text['label-call_recording_name'], $order_by, $order);
@@ -183,18 +216,18 @@
 	}
 	echo "	</td>\n";
 	echo "<tr>\n";
-	
+
 	if (is_array($result)) {
 		$x = 0;
 		foreach($result as $row) {
-			//if (permission_exists('recording_play') && $recording_file_path != '') {
+			//if (permission_exists('call_recording_play') && $recording_file_path != '') {
 			//	echo "<tr id='recording_progress_bar_".$row['call_recording_uuid']."' style='display: none3;'><td class='".$row_style[$c]." playback_progress_bar_background' style='padding: 0; border: none;' colspan='".((if_group("admin") || if_group("superadmin") || if_group("cdr")) ? ($col_count - 1) : $col_count)."'><span class='playback_progress_bar' id='recording_progress_".$row['call_recording_uuid']."'></span></td></tr>\n";
 			//}
 			if (permission_exists('call_recording_edit')) {
 				$tr_link = "href='call_recording_edit.php?id=".$row['call_recording_uuid']."'";
 			}
 			echo "<tr ".$tr_link.">\n";
-			echo "	<td valign='top' class='".$row_style[$c]." tr_link_void' style='text-align: center; padding: 3px 3px 0px 8px;'>\n";
+			echo "	<td valign='top' class='".$row_style[$c]." tr_link_void' style='align: center; padding: 3px 3px 0px 8px;'>\n";
 			echo "		<input type='checkbox' name=\"call_recordings[$x][checked]\" id='checkbox_".$x."' value='true' onclick=\"if (!this.checked) { document.getElementById('chk_all_".$x."').checked = false; }\">\n";
 			echo "		<input type='hidden' name=\"call_recordings[$x][call_recording_uuid]\" value='".$row['call_recording_uuid']."' />\n";
 			echo "	</td>\n";
@@ -204,14 +237,15 @@
 			//echo "			<source src=\"download.php?id=".$row['call_recording_uuid']."\" type=\"audio/wav\">\n";
 			//echo "		</audio>\n";
 			//echo "		<a href=\"download.php?id=".$row['call_recording_uuid']."&t=bin\">".$text['label-download']." ".$v_link_label_download."</a>\n";
-				
-			//if (permission_exists('recording_play')) {
-				echo 	"<audio id='recording_audio_".$row['call_recording_uuid']."' style='display: none;' preload='none' ontimeupdate=\"update_progress('".$row['call_recording_uuid']."')\" onended=\"recording_reset('".$row['call_recording_uuid']."');\" src=\"download.php?id=".$row['call_recording_uuid']."\" type='".$recording_type."'></audio>";
-				echo 	"<span id='recording_button_".$row['call_recording_uuid']."' onclick=\"recording_play('".$row['call_recording_uuid']."')\" title='".$text['label-play']." / ".$text['label-pause']."'>".$v_link_label_play."</span>";
-			//}
-			//if (permission_exists('recording_download')) {
-				echo 	"<a href=\"download.php?id=".$row['call_recording_uuid']."&t=bin\" title='".$text['label-download']."'>".$v_link_label_download."</a>";
-			//}
+			if (file_exists($row['call_recording_path'].'/'.$row['call_recording_name'])) {	
+				if (permission_exists('call_recording_play')) {
+					echo 	"<audio id='recording_audio_".$row['call_recording_uuid']."' style='display: none;' preload='none' ontimeupdate=\"update_progress('".$row['call_recording_uuid']."')\" onended=\"recording_reset('".$row['call_recording_uuid']."');\" src=\"download.php?id=".$row['call_recording_uuid']."\" type='".$recording_type."'></audio>";
+					echo 	"<span id='recording_button_".$row['call_recording_uuid']."' onclick=\"recording_play('".$row['call_recording_uuid']."')\" title='".$text['label-play']." / ".$text['label-pause']."'>".$v_link_label_play."</span>";
+				}
+				if (permission_exists('call_recording_download')) {
+					echo 	"<a href=\"download.php?id=".$row['call_recording_uuid']."&t=bin\" title='".$text['label-download']."'>".$v_link_label_download."</a>";
+				}
+			}
 			echo "	</td>\n";
 			//echo "	<td valign='top' class='".$row_style[$c]."' style=\"\">\n";
 			//echo "		<a href=\"download.php?id=".$row['call_recording_uuid']."&t=bin\">".$text['label-download']." ".$v_link_label_download."</a>\n";
